+++ Changelog.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+### Version 3.1.0 (2021-02-10)
+* Adding Collapse components
+* Adding Tab components
+* Changed link selection for component blocks and wysiwyg
+* Fix multi bug in the templates, scss and javascript
+
+
 ### Version 3.0.0 (2021-02-10)
 * Created new repository and new branch in to Github
 * Changed the project License to the MIT